Fix eosio symbol file and split tests into smaller files
