+++ .idea/markdown-navigator/profiles_settings.xml
@@ -1,0 +1,3 @@
+<component name="MarkdownNavigator.ProfileManager">
+  <settings default="" pdf-export="" plain-text-search-scope="Project Files" />
+</component>
