+++ build.sh
@@ -1,6 +1,12 @@
 #!/usr/bin/env bash
 set -eo pipefail
 
+NONINTERACTIVE=true
+PROCEED=true
+
+# Fix for broken eosio include file
+cp tests/symbol.hpp /usr/local/eosio/include/eosio/chain/
+
 printf "\t=========== Building eosdao.contracts ===========\n\n"
 RED='\033[0;31m'
 NC='\033[0m'
