--- contracts/custodian/compile.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-source ../_compiled_contracts/daccustodian/compile_all.sh
