--- contracts/custodian/daccustodian.clauses.md
@@ -1,5 +1,0 @@
-<h1 class="clause">ENTIRE AGREEMENT</h1>
-This contract contains the entire agreement of the parties, for all described actions, and there are no other promises or conditions in any other agreement whether oral or written concerning the subject matter of this Contract. This contract supersedes any prior written or oral agreements between the parties. 
-
-<h1 class="clause">BINDING CONSTITUTION</h1>
-All the the action descibed in this contract are subject to the EOSDAC consitution as held at http://eosdac.io. This includes, but is not limited to membership terms and conditions, dispute resolution and severability.
