+++ contracts/token/include/token/token.hpp
@@ -13,7 +13,6 @@
 
 #include <string>
 
-
 namespace eosdao {
 
     using namespace eosio;
