+++ contracts/voting/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_contract(voting voting ${CMAKE_CURRENT_SOURCE_DIR}/src/voting.cpp)
-find_path(LIBEOSDAC_INCLUDE_DIR NAMES libeosdac/libeosdac.hpp PATHS external NO_CMAKE_FIND_ROOT_PATH)
+find_path(LIBEOSDAC_INCLUDE_DIR NAMES libeosdac/libeosdac.hpp PATHS ${CMAKE_CURRENT_SOURCE_DIR}/../../external NO_CMAKE_FIND_ROOT_PATH)
 message(STATUS "libeosdac => ${LIBEOSDAC_INCLUDE_DIR}")
 
 if((NOT LIBEOSDAC_INCLUDE_DIR) OR (NOT EXISTS ${LIBEOSDAC_INCLUDE_DIR}))
-    message("Unable to find libeosdac ${LIBEOSDAC_INCLUDE_DIR} in voting contract.  Please make sure you have run git submidule update --init)")
+    message("Unable to find libeosdac ${LIBEOSDAC_INCLUDE_DIR} in voting contract.  Please make sure you have run git submodule update --init)")
 endif()
 
 target_include_directories(voting
