+++ contracts/voting/src/voting.cpp
@@ -56,13 +56,13 @@ namespace eosdao {
         auto dac = directory::dac_for_id(dac_id);
         eosio::name custodian_contract = dac.account_for_type(directory::types::CUSTODIAN);
 
+        print("notifying weight change to ", custodian_contract, "::weightobsv");
 //        eosio::action(
 //                eosio::permission_level{ get_self(), "notify"_n },
 //                custodian_contract, "weightobsv"_n,
 //                make_tuple(account_weights, dac.dac_id)
 //        ).send();
 
-        print("notifying weight change to ", custodian_contract, "::weightobsv");
     }
